third_party/WebKit/LayoutTests/inspector/profiler/heap-snapshot-statistics.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="heap-snapshot-test.js"></script>
<script>

function test()
{
    function createHeapSnapshot()
    {
        var builder = new InspectorTest.HeapSnapshotBuilder();
        var index = 0;
        for (type in InspectorTest.HeapNode.Type) {
            if (!InspectorTest.HeapNode.Type.hasOwnProperty(type)) continue;
            ++index;
            var size = index * Math.pow(10, index - 1);
            var node = new InspectorTest.HeapNode(type, size, InspectorTest.HeapNode.Type[type]);
            InspectorTest.addResult(type + " node size: " + size);
            builder.rootNode.linkNode(node, InspectorTest.HeapEdge.Type.internal, type + "Link");
        }
        return builder.generateSnapshot();
    }

    InspectorTest.runHeapSnapshotTestSuite([
        function testStatistics(next)
        {
            InspectorTest.addSniffer(WebInspector.HeapSnapshotView.prototype, "_gotStatistics", step1, true);
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, function() {});

            function step1(statistics)
            {
                InspectorTest.addResult(JSON.stringify(statistics));
                setTimeout(next, 0);
            }
        }
    ]);
}

</script>
</head>
<body onload="runTest()">
<p>
Tests Statistics view of detailed heap snapshots.
</p>
</body>
</html>
