third_party/WebKit/LayoutTests/inspector/profiler/heap-snapshot-statistics-expected.txt

Tests Statistics view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testStatistics
hidden node size: 1
array node size: 20
string node size: 300
object node size: 4000
code node size: 50000
closure node size: 600000
regexp node size: 7000000
number node size: 80000000
native node size: 900000000
{"total":987654321,"v8heap":87654321,"native":900000000,"code":50000,"arrays":20,"strings":300}

Profiler was disabled.
